create sale-report

+++ BackendAPI/Controllers/ProductSampleController.cs
@@ -19,12 +19,16 @@ namespace BackendAPI.Controllers
         private readonly IProductSampleService _productSampleService;
         private readonly IUnitOfWork _unitOfWork;
         private readonly IPhotoService _photoService;
+        private readonly IProductVersionService _productVersionService;
+        private readonly IProductColorProductService _productColorProductService;
 
-        public ProductSampleController(IProductSampleService productSampleService, IUnitOfWork unitOfWork, IPhotoService photoService)
+        public ProductSampleController(IProductSampleService productSampleService, IUnitOfWork unitOfWork, IPhotoService photoService, IProductVersionService productVersionService, IProductColorProductService productColorProductService)
         {
             _productSampleService = productSampleService;
             _unitOfWork = unitOfWork;
             _photoService = photoService;
+            _productVersionService = productVersionService;
+            _productColorProductService = productColorProductService;
         }
 
         [HttpGet]
@@ -62,6 +66,29 @@ namespace BackendAPI.Controllers
                 });
             }
         }
+        [HttpGet("get-all-photos-by-product-id/{id}")]
+        public async Task<IActionResult> GetAllPhotosByProductId(int id)
+        {
+            try
+            {
+
+                var photos = await _photoService.GetAllPhotosByProductId(id);
+                return Ok(new Response
+                {
+                    Data = photos,
+                    Success = true,
+
+                });
+            }
+            catch (Exception)
+            {
+                return StatusCode(StatusCodes.Status500InternalServerError, new Response
+                {
+                    Success = false,
+                    Errors = new[] { "Đã có lỗi xảy ra, vui lòng thử lại sau" }
+                });
+            }
+        }
         [HttpPost("upload-default-image")]
         public async Task<IActionResult> UploadProductSamppleDefaulImage([FromForm] UploadProductSampleDefaultImageRequest model)
         {
@@ -74,7 +101,7 @@ namespace BackendAPI.Controllers
                                                   .ToArray();
                     return BadRequest(new Response { Success = false, Errors = errors });
                 }
-                var productSample = await _productSampleService.GetProductSampleById(model.Id);
+                var productColorProduct = await _productColorProductService.GetProductColorProductById(model.Id);
                 string[] departmentArray = { ".png", ".jpeg", ".jpg", ".webp" };
                 string checkEntentions = Path.GetExtension(model.Image.FileName).ToLower();
 
@@ -88,14 +115,14 @@ namespace BackendAPI.Controllers
                 }
                 var file1 = Path.GetFileNameWithoutExtension(Path.GetRandomFileName())
             + Path.GetExtension(model.Image.FileName);
-                var fileTemp = Path.Combine("Uploads", "ProductSample", file1);
+                var fileTemp = Path.Combine("Uploads", "ProductColorProduct", file1);
 
                 using (var filestream = new FileStream(fileTemp, FileMode.Create))
                 {
                     await model.Image.CopyToAsync(filestream);
                 }
-                productSample.FileName = file1;
-                await _productSampleService.UpdateProductSample(productSample.Id, productSample);
+                productColorProduct.FileName = file1;
+                await _productColorProductService.UpdateProductColorProduct(productColorProduct.Id, productColorProduct);
                 await _unitOfWork.SaveChangesAsync();
                 return Ok(new Response
                 {
@@ -119,8 +146,8 @@ namespace BackendAPI.Controllers
         {
             try
             {
-                var findProductSample = await _productSampleService.GetProductSampleById(id);
-                if (findProductSample is null)
+                var findProductColorProduct = await _productColorProductService.GetProductColorProductById(id);
+                if (findProductColorProduct is null)
                 {
                     return BadRequest(new Response
                     {
@@ -129,10 +156,16 @@ namespace BackendAPI.Controllers
 
                     });
                 }
-                var filename = "Uploads/ProductSample/" + findProductSample.FileName;
-                System.IO.File.Delete(filename);
-                findProductSample.FileName = null;
-                await _unitOfWork.SaveChangesAsync();
+                var filename = "Uploads/ProductColorProduct/" + findProductColorProduct.FileName;
+                if (!string.IsNullOrEmpty(findProductColorProduct.FileName))
+                {
+                    System.IO.File.Delete(filename);
+
+                    findProductColorProduct.FileName = null;
+                    await _productColorProductService.UpdateProductColorProduct(findProductColorProduct.Id, findProductColorProduct);
+                    await _unitOfWork.SaveChangesAsync();
+                }
+
                 return Ok(new Response
                 {
                     Success = true,
@@ -181,7 +214,7 @@ namespace BackendAPI.Controllers
                         var file1 = Path.GetFileNameWithoutExtension(Path.GetRandomFileName())
                     + Path.GetExtension(f.FileName);
 
-                        var file = Path.Combine("Uploads", "ProductSample", file1);
+                        var file = Path.Combine("Uploads", "ProductColorProduct", file1);
 
                         using (var filestream = new FileStream(file, FileMode.Create))
                         {
@@ -190,7 +223,7 @@ namespace BackendAPI.Controllers
                         var photo = new Photo
                         {
                             FileName = file1,
-                            ProductSampleId = model.Id
+                            ProductColorProductId = model.Id
                         };
                         await _photoService.CreatePhoto(photo);
                         await _unitOfWork.SaveChangesAsync();
@@ -228,7 +261,108 @@ namespace BackendAPI.Controllers
 
                     });
                 }
-                var filename = "Uploads/ProductSample/" + findPhoto.FileName;
+                var filename = "Uploads/ProductColorProduct/" + findPhoto.FileName;
+                System.IO.File.Delete(filename);
+                await _photoService.DeletePhoto(findPhoto.Id);
+                await _unitOfWork.SaveChangesAsync();
+                return Ok(new Response
+                {
+                    Success = true,
+                    Message = "Xóa ảnh thành công"
+                });
+            }
+            catch (Exception)
+            {
+                return StatusCode(StatusCodes.Status500InternalServerError, new Response
+                {
+                    Success = false,
+                    Errors = new[] { "Đã có lỗi xảy ra, vui lòng thử lại sau" }
+                });
+                throw;
+            }
+
+        }
+
+
+
+        [HttpPost("upload-image-product-360")]
+        public async Task<IActionResult> UploadImageProduct360([FromForm] UploadImageProduct360Request model)
+        {
+            try
+            {
+                if (!ModelState.IsValid)
+                {
+                    var errors = ModelState.Values.SelectMany(v => v.Errors)
+                                                  .Select(e => e.ErrorMessage)
+                                                  .ToArray();
+                    return BadRequest(new Response { Success = false, Errors = errors });
+                }
+
+                if (model.Images.Count > 0)
+                {
+                    string[] departmentArray = { ".png", ".jpeg", ".jpg", ".webp" };
+                    foreach (var f in model.Images)
+                    {
+                        string checkEntentions = Path.GetExtension(f.FileName).ToLower();
+
+                        if (departmentArray.Contains(checkEntentions) == false)
+                        {
+                            return BadRequest(new Response
+                            {
+                                Success = false,
+                                Message = "File không đúng định dạng",
+                            });
+                        }
+                        var file1 = f.FileName;
+
+                        var file = Path.Combine("Uploads", "ImageProduct360", file1);
+
+                        using (var filestream = new FileStream(file, FileMode.Create))
+                        {
+                            await f.CopyToAsync(filestream);
+                        }
+                        var photo = new Photo
+                        {
+                            FileName = file1,
+                            ProductId = model.Id
+                        };
+                        await _photoService.CreatePhoto(photo);
+                        await _unitOfWork.SaveChangesAsync();
+                    }
+                }
+                return Ok(new Response
+                {
+                    Success = true,
+                    Message = "Upload thành công"
+                });
+            }
+            catch (Exception)
+            {
+                return StatusCode(StatusCodes.Status500InternalServerError, new Response
+                {
+                    Success = false,
+                    Errors = new[] { "Đã có lỗi xảy ra, vui lòng thử lại sau" }
+                });
+                throw;
+            }
+
+        }
+        [HttpDelete("delete-image-product-360/{id}")]
+        public async Task<IActionResult> DeleteImageProduct360(int id)
+        {
+            try
+            {
+                var findPhoto = await _photoService.GetPhotoById(id);
+                if (findPhoto is null)
+                {
+                    return BadRequest(new Response
+                    {
+                        Success = false,
+                        Errors = new[] { "Không tìm thấy" }
+
+                    });
+                }
+                var filename = "Uploads/ImageProduct360/" + findPhoto.FileName;
                 System.IO.File.Delete(filename);
                 await _photoService.DeletePhoto(findPhoto.Id);
                 await _unitOfWork.SaveChangesAsync();
